moving uuid code into runtime library
moving test_uuid to test_runtime
