+++ source/library/runtime/CMakeLists.txt
@@ -7,9 +7,25 @@ add_library(potato::runtime ALIAS runtime)
 include(up_set_common_properties)
 up_set_common_properties(runtime)
 
+# setup uuid libs for different oses
+if (WIN32)
+    set(UP_UUID_LIBRARY "rpcrt4.lib")
+elseif(APPLE)
+    message(FATAL_ERROR "MacOS uuid lib not supported yet.")
+else()
+    find_package(libuuid REQUIRED)
+    if (NOT LIBUUID_FOUND)
+        message(FATAL_ERROR "You might need to run 'sudo apt-get install uuid-dev' or similar")
+    endif()
+    set(UP_UUID_LIBRARY ${LIBUUID_LIBRARY})
+    target_include_directories(foundation PRIVATE ${LIBUUID_INCLUDE_DIR})
+    target_compile_definitions(foundation PRIVATE GUID_LIBUUID)
+endif()
+
 target_link_libraries(runtime
     PUBLIC
         foundation
+        ${UP_UUID_LIBRARY}
 )
 
 target_compile_definitions(runtime
