+++ source/library/runtime/tests/test_uuid.cpp
@@ -1,4 +1,4 @@
-#include "potato/foundation/uuid.h"
+#include "potato/runtime/uuid.h"
 #include <doctest/doctest.h>
 #include <array>
